several memory leaks and potential reasons for a crash removed (thanks to loigu)
git-svn-id: https://pupnp.svn.sourceforge.net/svnroot/pupnp/trunk@13 119443c7-1b9e-41f8-b6fc-b9c35fce742c

+++ libupnp/upnp/src/inc/upnpapi.h
@@ -57,7 +57,7 @@
 #define DEFAULT_SOAP_CONTENT_LENGTH 16000
 #define MAX_SOAP_CONTENT_LENGTH 32000
 
-extern int g_maxContentLength;
+extern size_t g_maxContentLength;
 
 // 30-second timeout
 #define UPNP_TIMEOUT	30
